requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
pytest>=7.0.0
pytest-cov>=4.0.0
numpy>=1.22.0
sentence-transformers>=2.2.0